refactor: 이름 객체 분리
